share cred parsing between opensearch and ES. Surfaced some config params, but not all
