+++ SearchProviders/preloaded.json
@@ -312,7 +312,7 @@
         "response_mappings": "",
         "result_mappings": "url=_source.url,date_published=_source.date_published,author=_source.author,title=_source.subject,body=_source.content,_source.to,NO_PAYLOAD",
         "results_per_query": 10,
-        "credentials": "'admin','password'",
+        "credentials": "verify_certs=[True|False],ca_certs=/path/to/cert/file.crt,username:password",
         "tags": [
             "Enron",
             "OpenSearch",
